+++ bits/01_version.js
@@ -1,1 +1,1 @@
-XLSX.version = '0.18.10';
+XLSX.version = '0.18.11';